requirements.txt
datasets
pandas
nltk
scikit-learn
tensorflow-macos
tensorflow-metal
